Added input pooling to the game.
Before you could lose inputs if you were fast enough to issue a second input before the game loop had time to tick over. This was occuring because the logic was tick dependant and only checking the last issued input command to the snakes current direction. Input pooling fixes this by checking the last issued input command to the previous one. It will only pool two commands how ever, this is to stop the snake from ghosting if a very large amount of inputs are issued before the game ticks over.